Restructured api models, clients, and endpoints into api/ for future scalability
