+++ src/WalletWave/api/models/gmgn/token_info.py
