+++ src/WalletWave/api/models/gmgn/wallet_info.py
