+++ src/WalletWave/api/models/gmgn/wallets.py
